feat!: Implement POST beatmapset/{id:int}/favourited to change beatmap favourited status; feat!: Use `{id:int}/{mode}` and `self/{mode}` routes to get UserWithStatsResponse; feat!: Use body instead of queries to update users friendship status; feat: Add missing swagger data for `/leaderboard` endpoint; fix: Fix some wrong ProducesResponseType types;
ref: CalculateBeatmapPerformance uses GameMode type for gamemode;

+++ Sunrise.Server/Services/ScoreService.cs
@@ -207,7 +207,7 @@ public class ScoreService(BeatmapService beatmapService, DatabaseService databas
         // Mods can change difficulty rating, important to recalculate it for right medal unlocking
         if ((int)score.GameMode != beatmap.ModeInt || (int)score.Mods > 0)
         {
-            var recalculateBeatmapResult = await calculatorService.CalculateBeatmapPerformance(session, score.BeatmapId, (int)score.GameMode, score.Mods);
+            var recalculateBeatmapResult = await calculatorService.CalculateBeatmapPerformance(session, score.BeatmapId, score.GameMode, score.Mods);
 
             if (recalculateBeatmapResult.IsFailure)
             {
